requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.28.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
docx2txt==0.8
entrypoints==0.4
etelemetry==0.3.1
exceptiongroup==1.2.1
executing==2.0.1
Faker==25.0.1
favicon==0.7.0
filelock==3.14.0
fitz==0.0.1.dev2
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jiter==0.4.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
looseversion==1.3.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openai==1.25.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pillow==10.3.0
pip-chill==1.0.3
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
prov==2.0.0
psutil==6.0.0
pure-eval==0.2.2
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
pydot==2.0.0
Pygments==2.17.2
pymdown-extensions==10.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
rdflib==7.0.0
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
st-annotated-text==4.0.1
stack-data==0.6.3
streamlit>=1.34.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-tags==1.2.8
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
wcwidth==0.2.13
yarl==1.9.4